h2 local impl

+++ src/main/java/com/cowell/core/Queue.java
@@ -9,9 +9,9 @@ public interface Queue<T extends QueueElement> {
 
     void setName(String name);
 
-    int getCapacity();
+    long getCapacity();
 
-    int size();
+    long size();
 
     boolean offer(T element);
